server/requirements.txt
# Qwen3-VL API服务器依赖

# Web框架
flask>=2.0.0
flask-cors>=3.0.0

# 配置文件解析
pyyaml>=6.0

# 深度学习框架
torch>=2.0.0
transformers>=4.30.0

# 模型加载
modelscope>=1.0.0

# HTTP请求（如果需要）
requests>=2.28.0

# 定时任务调度（用于记忆训练）
apscheduler>=3.10.0

# 文件处理（PDF/DOCX文本和图片提取）
pdfminer.six>=20221105  # PDF文本提取
pdf2image>=1.16.0  # PDF转图片（需要系统安装poppler-utils）
python-docx>=0.8.11  # DOCX文件处理
pytesseract>=0.3.10  # OCR（可选，需要系统安装tesseract）
beautifulsoup4>=4.11.0  # 网页内容提取（可选）

